Fixed some 64-bit related compiler warnings.

+++ src/audio/sfx_manager.cpp
@@ -502,7 +502,7 @@ void SFXManager::loadSfx()
     delete root;
 
     // Now load them in parallel
-    const int max = m_all_sfx_types.size();
+    const int max = (int)m_all_sfx_types.size();
     SFXBuffer **array = new SFXBuffer *[max];
     i = 0;
